test/LayoutTests/Panel/PageURLChanges.html

<html>
<body>
<h1>Feature Test Chnage project</h1>
</body>
<script src="chrome-extension://klmlfkibgfifmkanocmdenpieghpgifl/ChannelPlate/ChannelPlate.js"></script>
<script src="chrome-extension://klmlfkibgfifmkanocmdenpieghpgifl/ChannelPlate/RemoteMethodCall.js"></script>
<script src="chrome-extension://klmlfkibgfifmkanocmdenpieghpgifl/AsyncMachine.js"></script>
<script src="../../../DebugLogger.js"></script>
<script src="../PatientSelector.js"></script>
<script src="../panel-test.js"></script>

<script>

window.addEventListener('load', function() {
  testScript();
});

function test_lastChange() {
  console.log("------------test begins------------");
  AsyncMachine.pushOp(evaluateInPage, 'window.location="http://localhost:8686/test/qpd.html"');
  AsyncMachine.pushOp(openQuerypointPanel);
  AsyncMachine.pushOp(openSourceFile, 'demo.js');
    var objPropSelector = [
      {type: 'span.cm-variable', text:'bar'},  // call in onClick
      {type: 'span.cm-variable', text:'bar'},  // func decl
      {type: 'span.cm-property', text: 'prop'},  // obj.prop = 0
      {type: 'span.cm-property', text: 'prop'}   // prop in obj.prop = 0
    ];
  AsyncMachine.pushOp(selectTokenInSource, objPropSelector);
  AsyncMachine.pushOp(verifyTokenView, 'obj.prop');
  AsyncMachine.pushOp(clickQPOperation, 'lastChange');
  AsyncMachine.pushOp(whenSelectorAll, '.currentLoadNumber', '1');
  AsyncMachine.pushOp(evaluateInPage, 'document.querySelector("#myButton").dispatchEvent(new MouseEvent("click"))');
  AsyncMachine.pushOp(whenSelectorAll, 'td.traceValue', '0');
  AsyncMachine.pushOp(extractText, 'td.traceValue');
  // Change the viewed page
  AsyncMachine.pushOp(evaluateInPage, 'window.location="http://localhost:8686/test/switch.html"');
  // Wait for the auto-reload
  AsyncMachine.pushOp(whenSelectorAll, '.currentLoadNumber', '1');
  // The new html should be available
  AsyncMachine.pushOp(openSourceFile, 'switch.html');
  // The qp runtime should still be loaded
  AsyncMachine.pushOp(evaluateInPage, 'window.__qp_reloads', function(text) {
    return 'window.__qp_reloads:' + (text ? text : "no text");
  });
  // The trace should be gone
  AsyncMachine.pushOp(extractText, 'td.traceValue');
  // The onPanelOpen dialog should be up
  AsyncMachine.pushOp(clickSelector, '.initialOptions_reloadWithoutQP', '', function(text) {
    return "Pressed Reload without QP runtime";
  });
  // Wait for the reload.
  AsyncMachine.pushOp(whenSelectorAll, '.runtimeStatus', 'QP runtime is not installed.');
    // The qp runtime should not be loaded
  AsyncMachine.pushOp(evaluateInPage, 'window.__qp_reloads', function(text) {
    return 'window.__qp_reloads:' + (text ? text : "no text");
  });
  AsyncMachine.runOp();
}

var testScript = test_lastChange;
</script>
</html>
