test/LayoutTests/Panel/PageURLChanges-expected.txt

http://localhost:8686/test/qpd.html
prop in     obj.prop = 0;
true
1|0
http://localhost:8686/test/switch.html
window.__qp_reloads:1
window.__qp_reloads:no text
